requirements.txt
pandas
numpy
tqdm
python-chess
keras
tensorflow